==> notebooks/03_Feature_Engineering.ipynb <==
# Step 3 - Feature Engineering

Normalize features to ensure they contribute equally to the model. This process can help improve the performance and convergence of various machine learning algorithms.

After this, we will perform data augmentation using SMOTE technique.

import pandas as pd
from sklearn.preprocessing import StandardScaler
from imblearn.over_sampling import SMOTE

## 3.0 - Load the Data

previous_years_df = pd.read_csv('../data/02_previous_years_not_null.csv')
present_year_df = pd.read_csv('../data/04_present_year_not_null.csv')

## 3.1 - Scaling the Data

Since our features do not have the same units or scale, it's generally better to use StandardScaler in this case. StandardScaler standardizes features to have zero mean and unit variance, which is often more suitable when dealing with features of varying scales or units.

Besides, we will also encode the class 'neg' as 0 and class 'pos' as 1 for further use.


# Function to scale the data
def scaling(data):
    
    """
    Scales the features of a dataset using StandardScaler and returns the scaled dataset.
    """

    # Separate the target and feature columns
    X = data.drop('class', axis=1)  # Features
    y = data['class']               # Target

    # Initialize the StandardScaler
    scaler = StandardScaler()
    X_scaled = scaler.fit_transform(X)

    # Convert the scaled features back to a DataFrame
    X_scaled_df = pd.DataFrame(X_scaled, columns=X.columns)
    scaled_data = pd.concat([X_scaled_df, y.reset_index(drop=True)], axis=1)

    # Map target values to numeric
    scaled_data['class'] = scaled_data['class'].map({'neg': 0, 'pos': 1})

    return scaled_data

**Previous Years**

previous_scaled = scaling(previous_years_df)
previous_scaled.head(5)

**Present Year:**

present_scaled = scaling(present_year_df)
present_scaled.head(5)

## 3.2 - Saving Original Data Scaled

previous_scaled.to_csv('../data/05_previous_years_scaled.csv', index=False)
present_scaled.to_csv('../data/06_present_year_scaled.csv', index=False)

## 3.3 - Data Augmentation

Apply SMOTE (Synthetic Minority Over-sampling Technique) to balance the class distribution by generating synthetic samples for the minority class.

It should be applied only to the training data to prevent data leakage and ensure that the test data remains representative of the real-world scenario. It will be explain in detaild in the 5th notebook that the training data here is the Previous Years data.

# Separate target and features for training data
y_train = previous_scaled['class']
X_train = previous_scaled.drop('class', axis=1)

# Apply SMOTE to training data
smote = SMOTE(random_state=42)
X_train_resampled, y_train_resampled = smote.fit_resample(X_train, y_train)

# Count the occurrences of each class in the resampled target variable
class_counts = pd.Series(y_train_resampled).value_counts()

print("Training Resampled features shape:", X_train_resampled.shape)
print("Training Resampled target shape:", y_train_resampled.shape)
print(f"\nClass Count: {class_counts}")

## 3.4 - Saving Resampled (Augmented) Data

# Create a DataFrame from the resampled data
X_train_resampled_df = pd.DataFrame(X_train_resampled, columns=X_train.columns)
y_train_resampled_df = pd.DataFrame(y_train_resampled, columns=['class'])

# Combine features and target into a single DataFrame
SMOTEd_df = pd.concat([X_train_resampled_df, y_train_resampled_df], axis=1)

SMOTEd_df.to_csv('../data/07_previous_years_SMOTEd.csv', index=False)

## Note:

If I have more time available for this work, I will also apply feature selection as part of the feature engineering process.

One simple way to do this, as demonstrated in `00_Challenge_Activities.ipynb`, is to use feature importance from a basic tree-based model (e.g., Random Forest).

After this step, I will have a new subset of data, which I can then apply the SMOTE technique and use for model training and testing.
